+++ lib/tree_fuzzer.ml
@@ -19,43 +19,53 @@ let fuzzer (max_depth: int) (values: (element,string) Hashtbl.t option) (goal: e
             else (Hashtbl.hash r2.right_part) - (Hashtbl.hash r1.right_part) (* to be deterministic we want to compare any couple *)
         end in
 
-    (* not tail-recursive ! *)
-    let rec fuzzer_minimize (goal_rules: rule list) (used_rules: rule list) (e: element) : parse_tree =
-        if Option.map (fun v -> Hashtbl.mem v e) values = Some true then
-            Leaf (Terminal (Hashtbl.find (Option.get values) e))
-        else if is_terminal e then
-            Leaf e
+    (* like map, but stop whenever the fuzzing fails in one branch *)
+    (* not tail-recursive *)
+    let rec lazy_map (f: element -> parse_tree) (l: element list) : parse_tree list = match l with
+        | [] -> []
+        | t::q -> begin match f t with
+                | Error -> [Error]
+                | x -> x::(lazy_map f q) end in
+
+    let rec fuzzer_minimize (goal_rules: rule list) (used_symbols: element list) (used_rules: rule list) (root: element) : parse_tree =
+        if Option.map (fun v -> Hashtbl.mem v root) values = Some true then
+            Leaf (Terminal (Hashtbl.find (Option.get values) root))
+        else if is_terminal root then
+            Leaf root
         else if goal_rules <> [] then begin
             let r = List.hd goal_rules in
             if List.tl goal_rules = [] then
-                Node (List.map (fuzzer_minimize [] (r::used_rules)) r.right_part)
+                (* don't update used_symbols as the path to goal may require producing a tree with two identical symbols in the same branch *)
+                Node (List.map (fuzzer_minimize [] used_symbols (r::used_rules)) r.right_part)
             else begin
                 let q = List.tl goal_rules in
-                let e = (List.hd q).left_symbol in
-                let first = List.find ((=) e) r.right_part in
+                let next_sym = (List.hd q).left_symbol in
+                let first = List.find ((=) next_sym) r.right_part in
                 (* the PHYSICAL equality is not an error : we want the first, not all of them *)
-                Node (List.map (fun e -> fuzzer_minimize (if e == first then q else []) (r::used_rules) e) r.right_part)
+                Node (List.map (fun e -> fuzzer_minimize (if e == first then q else []) used_symbols (r::used_rules) e) r.right_part)
             end
-        end else begin
-            let possible_rules = List.filter (fun r -> r.left_symbol = e && not (List.mem r used_rules)) g.rules in
-            if Hashtbl.mem nonrec_rules e then
-                let r = Hashtbl.find nonrec_rules e in
-                Node (List.map (fuzzer_minimize goal_rules (r::used_rules)) r.right_part)
-            else if possible_rules = [] then
-                Error
+        end else if List.mem root used_symbols then Error
+        else begin
+            (* goal_rules is empty *)
+            match Hashtbl.find_opt nonrec_rules root with
+            | Some r -> Node (List.map (fuzzer_minimize [] (root::used_symbols) (r::used_rules)) r.right_part) (* a good rule is already known *)
+            | None ->
+                (* get the rules root -> ... that are not used and with unused symbols in the RHS *)
+                let possible_rules = List.sort compare_rule (List.filter (fun r -> r.left_symbol = root && not (List.mem r used_rules) && not (List.exists (fun e -> List.mem e used_symbols) r.right_part)) g.rules) in
+                use_rule possible_rules (root::used_symbols) used_rules root
+        end
+
+    and use_rule (possible_rules: rule list) (used_symbols: element list) (used_rules: rule list) (root: element) : parse_tree = match possible_rules with
+        | [] -> Error
+        | r::q ->
+            let trees = lazy_map (fuzzer_minimize [] used_symbols (r::used_rules)) r.right_part in
+            if List.exists ((=) Error) trees then
+                (use_rule [@tailcall]) q used_symbols used_rules root
             else begin
-                let r = List.hd (List.sort_uniq compare_rule possible_rules) in (* rules with least nonterminals *)
-                let new_used_rules = r::used_rules in (* rules in the exploding area are ignored for duplication purpose *)
-                let trees = List.map (fuzzer_minimize goal_rules (new_used_rules)) r.right_part in
-                if List.exists ((=) Error) trees then
-                    (fuzzer_minimize [@tailcall]) goal_rules (new_used_rules) e (* restart with a different rule *)
-                else begin
-                    (* we need to verify the hashtable after the recursive call *)
-                    if not (Hashtbl.mem nonrec_rules e) then Hashtbl.add nonrec_rules e r;
-                    Node trees
-                end
-            end
-        end in
+                (* we need to verify the hashtable after the recursive call. This rule is safe since the recursive call succeeded! *)
+                Hashtbl.replace nonrec_rules root r;
+                Node trees
+            end in
 
     (* not tail-recursive ! but the max depth is controlled *)
     let rec fuzzer_explode (depth: int) (e: element) : parse_tree =
@@ -65,7 +75,8 @@ let fuzzer (max_depth: int) (values: (element,string) Hashtbl.t option) (goal: e
         end else if is_terminal e then
             Leaf e
         else if depth >= max_depth then
-            fuzzer_minimize [] [] e
+            fuzzer_minimize [] [] [] e
+            (*fuzzer_minimize [] [] e*)
         else begin
             let possible_rules = List.filter (fun r -> r.left_symbol = e) g.rules in
             let r = List.nth possible_rules (Random.int (List.length possible_rules)) in (* random rule *)
@@ -80,7 +91,9 @@ let fuzzer (max_depth: int) (values: (element,string) Hashtbl.t option) (goal: e
         | t::_ when not (List.mem t seen) -> true
         | _::q -> (has_new [@tailcall]) seen q in
 
-    let rec find_path_to_goal_aux (seen: element list) (queue : (part * rule list) list) : rule list = match queue with
+    let rec find_path_to_goal_aux (seen: element list) (queue : (part * rule list) list) : rule list =
+        assert (goal <> None);
+        match queue with
         | [] -> assert false (* we know there is a path since the goal is reachable ! *)
         | (form,path)::_ when List.mem (Option.get goal) form -> List.rev path
         | (form,_)::q when not (has_new seen form) -> (find_path_to_goal_aux [@tailcall]) seen q
@@ -96,7 +109,12 @@ let fuzzer (max_depth: int) (values: (element,string) Hashtbl.t option) (goal: e
             if List.exists ((=) None) trees then None
             else Some (trees |> List.map Option.get |> List.flatten) in
 
-    if verbose then print_endline "Fuzzing";
-    let tree = if Option.is_some goal && is_reachable g (Option.get goal) g.axiom then fuzzer_minimize (find_path_to_goal ()) [] g.axiom
-    else (fuzzer_explode 0 g.axiom) in
-    part_of_tree tree
+    part_of_tree (
+        if Option.is_some goal && is_reachable g (Option.get goal) g.axiom then begin
+            if verbose then print_endline "Fuzzing with goal";
+            fuzzer_minimize (find_path_to_goal ()) [] [] g.axiom
+(*        fuzzer_minimize (find_path_to_goal ()) [] g.axiom*)
+        end else begin
+            if verbose then print_endline "Fuzzing";
+            fuzzer_explode 0 g.axiom
+        end)